Time: 23 ms (58.14%), Space: 70.4 MB (17.34%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -25,7 +25,7 @@ public:
     int minDays(vector<int>& bloomDay, int m, int k) {
         
         int n = bloomDay.size();
-        if(m*k > n) return -1;
+        if((long) m*k > n) return -1;
 
         int s = *min_element(bloomDay.begin(), bloomDay.end());
         int e = *max_element(bloomDay.begin(), bloomDay.end());
